Re-usable REPL buffer
The menu option will refocus the same REPL buffer rather than make a new one, if the old buffer is still open.

The 'open_repl' function is exposed for key-binds.
With no arguments, works as the menu option above.  With a truthy argument, creates a new buffer each time, the old behavior.  With false as an argument, it will refocus an existing REPL if one exists, but will not create a new one.

Key-binds for REPLs are reloaded after Lua state reset.